refactor: convert cron_manager builtin to package pattern (#115) (#116)
